Refactor ConfiguratorIntegrationService
Refactor ConfiguratorIntegrationService.ts into smaller files to improve maintainability.
